Jan 2025 release

+++ vnncomp_scripts/prepare_instance.sh
@@ -5,7 +5,6 @@ VERSION_STRING=v1
 
 if [[ -z "${VNNCOMP_PYTHON_PATH}" ]]; then
 	VNNCOMP_PYTHON_PATH=/home/ubuntu/miniconda/envs/alpha-beta-crown/bin
-	VNNCOMP_PYTHON_LEGACY_PATH=/home/ubuntu/miniconda/envs/alpha-beta-crown-2022/bin
 fi
 echo $VNNCOMP_PYTHON_PATH
 
@@ -27,48 +26,31 @@ echo TOOL_DIR is $TOOL_DIR
 export PYTHONPATH=${TOOL_DIR}
 export OMP_NUM_THREADS=1
 
-# kill any zombie processes
-killall -q python
-killall -q python3
-killall -q get_cuts
-killall -q -9 python
-killall -q -9 python3
-killall -q -9 get_cuts
-sleep 3
-# Reset GPU, make sure nothing is running.
-(sudo -n rmmod nvidia_uvm; sudo -n rmmod nvidia_drm; sudo -n rmmod nvidia_modeset; sudo -n nvidia-smi -e 0; sudo -n nvidia-smi -pm 0; sudo -n nvidia-smi -r -i 0) > /dev/null
-sudo -n modprobe nvidia_uvm; sudo -n nvidia-smi -pm 1
-# Make sure GPU shows up.
-nvidia-smi
+# If ABCROWN_TEST_RUN is 1, then we skip the extra steps to make the run faster, for debugging propose.
+if [[ "${ABCROWN_TEST_RUN}" -ne "1" ]]; then
 
-# Convert MaxPool in vgg16-7.onnx into ReLU
-# Dependency: pip install git+https://github.com/dlshriver/DNNV.git@develop
-if [ "$CATEGORY" == "vggnet16_2022" ]; then
-	if [ -f "$ONNX_FILE.original" ]; then
-		echo 'vgg16-7.onnx previously converted'
-	else
-	  echo 'vgg16-7.onnx converting...'
-		${VNNCOMP_PYTHON_LEGACY_PATH}/python ${TOOL_DIR}/vnncomp_scripts/maxpool_to_relu.py $ONNX_FILE
-		cp $ONNX_FILE $ONNX_FILE.original
-		cp output.onnx $ONNX_FILE
-	fi
+	# kill any zombie processes
+	killall -q python
+	killall -q python3
+	killall -q get_cuts
+	killall -q -9 python
+	killall -q -9 python3
+	killall -q -9 get_cuts
+	sleep 3
+	# Reset GPU, make sure nothing is running.
+	(sudo -n rmmod nvidia_uvm; sudo -n rmmod nvidia_drm; sudo -n rmmod nvidia_modeset; sudo -n nvidia-smi -e 0; sudo -n nvidia-smi -pm 0; sudo -n nvidia-smi -r -i 0) > /dev/null
+	sudo -n modprobe nvidia_uvm; sudo -n nvidia-smi -pm 1
+	# Make sure GPU shows up.
+	nvidia-smi
+
 fi
-if [ "$CATEGORY" == "vggnet16" ]; then
-	if [ -f "$ONNX_FILE.original" ]; then
-		echo 'vgg16-7.onnx previously converted'
-	else
-	  echo 'vgg16-7.onnx converting...'
-		${VNNCOMP_PYTHON_LEGACY_PATH}/python ${TOOL_DIR}/vnncomp_scripts/maxpool_to_relu.py $ONNX_FILE
-		cp $ONNX_FILE $ONNX_FILE.original
-		cp output.onnx $ONNX_FILE
-	fi
-fi
-if [ "$CATEGORY" == "cgan" ]; then
+
+if [[ "$CATEGORY" == "cgan" || "$CATEGORY" == "cgan_2023" ]]; then
 	if [ -f "$ONNX_FILE.original" ]; then
 		echo 'cgan models previously simplified'
 	else
-	  echo 'cgan simplifying...'
-	  ${VNNCOMP_PYTHON_PATH}/onnxsim "$ONNX_FILE" output.onnx
+		echo 'cgan simplifying...'
+		${VNNCOMP_PYTHON_PATH}/onnxsim "$ONNX_FILE" output.onnx
 		cp $ONNX_FILE $ONNX_FILE.original
 		cp output.onnx $ONNX_FILE
 	fi
@@ -79,30 +61,23 @@ echo
 echo "Running warmup..."
 echo
 temp_file=$(mktemp)
-if [ "$CATEGORY" == "nn4sys" ]; then
-	prepare_timeout=275
-elif [ "$CATEGORY" == "vggnet16_2022" ]; then
-	prepare_timeout=90
-elif [ "$CATEGORY" == "tllverifybench" ]; then
-	prepare_timeout=90
-elif [ "$CATEGORY" == "collins_yolo_robustness" ]; then
-	prepare_timeout=180
-else
-	prepare_timeout=45
-fi
-echo "Preparation time is roughly ${prepare_timeout} seconds for $CATEGORY"
-timeout -k 5 ${prepare_timeout} ${VNNCOMP_PYTHON_PATH}/python ${TOOL_DIR}/complete_verifier/vnncomp_main.py "$CATEGORY" "$ONNX_FILE" "$VNNLIB_FILE" "$temp_file" 1 > /dev/null
-rm ${temp_file}
+prepare_timeout=300
+echo "Preparation time budget is ${prepare_timeout} seconds for $CATEGORY"
+if [[ "${ABCROWN_TEST_RUN}" -ne "1" ]]; then
+	timeout -k 5 ${prepare_timeout} ${VNNCOMP_PYTHON_PATH}/python ${TOOL_DIR}/complete_verifier/vnncomp_main.py "$CATEGORY" "$ONNX_FILE" "$VNNLIB_FILE" "$temp_file" 1 --PREPARE >/dev/null
+	rm ${temp_file}
 
-# kill any remaining python processes.
-killall -q python
-killall -q python3
-killall -q get_cuts
-sleep 1
-killall -q -9 python
-killall -q -9 python3
-killall -q -9 get_cuts
-sleep 3
+	# kill any remaining python processes.
+	killall -q python
+	killall -q python3
+	killall -q get_cuts
+	sleep 1
+	killall -q -9 python
+	killall -q -9 python3
+	killall -q -9 get_cuts
+	sleep 3
+fi
+
 echo "Preparation finished."
 
 # script returns a 0 exit code if successful. If you want to skip a benchmark category you can return non-zero.
